fix: ORB errors from incorrect query parsing
urls passed to one of the image hosts, `wsrv.nl`, must have their query
strings (if any) properly encoded, otherwise the compression fails. in
simple compression mode, due to limitations with MV3's
`declarativeNetRequest`, we cannot do the encoding so the next best
thing is to use the original image as the default.